+++ src/PhpSpreadsheet/Shared/OLERead.php
@@ -326,10 +326,7 @@ class OLERead
      */
     private static function getInt4d($data, $pos)
     {
-        if (trim($data) == '') {
-            // No data provided
-            throw new ReaderException('Parameter data is empty.');
-        } elseif ($pos < 0) {
+        if ($pos < 0) {
             // Invalid position
             throw new ReaderException('Parameter pos=' . $pos . ' is invalid.');
         }